cpc::front::parser
Function proper_repl_instruction

pub fn proper_repl_instruction<'a>(
    file: &'a str,
    input: &'a str,
) -> Result<ReplInstr, String>

Parser for Repl Instructions

This parser reads a REPL instruction by consuming the entire input. Note that this ignores whitespaces at the beginning of the input as well as at the end.