cpc::kernel
Module nbe

Normalization Module Based on Normalization-by-Evaluation (NbE)

This module provides normalization methods (nbe and nbe_typ) for a well-typed term (and well-formed type). Note that, for well-typed term, normalization should never fail, any errorneous cases in this module call panic!.

Actual implementation lives in trait implementations of Exp. Specifically, one can check its nbe method and eval method.

Traits

Readback 🔒

Readback from Domain Types (DomNorm and DomNeut)

Functions

build_initial_env 🔒

inconsistent_panic 🔒

Helper for Inconsistency Case

non_neutral_exp_panic 🔒

Helper for Non-neutral Expression Case when Expecting Neutral

non_type_exp_panic 🔒

Helper for Non-type Expression Case when Expecting Type

reduce_app 🔒

Reduce Domain-value Application

Type Aliases

Idents 🔒