cpc::kernel
Module nbe

Normalization Module Based on Normalization-by-Evaluation (NbE)

This module provides normalization methods (nbe and nbe_typ) for a well-typed term (and well-formed type). Note that, for well-typed term, normalization should never fail, any errorneous cases in this module call panic!.

Actual implementation lives in trait implementations of Exp. Specifically, one can check its nbe method and eval method.

Modules

domain 🔒

Semantic Domain For Evaluation

evaluation 🔒

Evaluation into Domain

panic 🔒

readback 🔒

Readback from Domain

Functions

build_initial_env 🔒

env_into_idents 🔒

subst_nbe_typ

Eager Substitution-Normalization for Types