cpc::kernel::nbe::readback
Trait Readback

pub trait Readback {
    type NormalizedSyntax;

    // Required method
    fn readback(self, ids: &mut HashSet<Ident>) -> Self::NormalizedSyntax;
}

Main Trait for Readback

Required Associated Types

type NormalizedSyntax

Required Methods

fn readback(self, ids: &mut HashSet<Ident>) -> Self::NormalizedSyntax

This method takes the ownership of self as we also want to take the ownership of Dom as a domain type in domain values.